Refactor lpr into real time data processor
